+++ frontend/nextjs/next.config.js
@@ -2,14 +2,6 @@
 const nextConfig = {
   reactStrictMode: true,
   output: 'standalone',
-  images: {
-    remotePatterns: [
-      {
-        protocol: 'https',
-        hostname: '**',
-      },
-    ],
-  },
 }
 
 module.exports = nextConfig